% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxlib_close_file.R
\name{pxlib_close_file}
\alias{pxlib_close_file}
\title{Close a Paradox database file}
\usage{
pxlib_close_file(pxdoc)
}
\arguments{
\item{pxdoc}{An external pointer of class 'pxdoc_t' obtained from \code{pxlib_open_file()}.}
}
\value{
Invisible \code{NULL}.
}
\description{
This function explicitly closes a Paradox database file associated with
a \code{pxdoc_t} external pointer and releases its resources.
}
