% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pxlib_open_file.R
\name{pxlib_open_file}
\alias{pxlib_open_file}
\title{Open a Paradox database file}
\usage{
pxlib_open_file(path)
}
\arguments{
\item{path}{A character string specifying the path to the Paradox (.db) file.}
}
\value{
An external pointer of class 'pxdoc_t' if the file is successfully
opened, or \code{NULL} if an error occurs.
}
\description{
This function opens a Paradox database file. It will also automatically
search for and attach an associated BLOB file (.mb) in a case-insensitive
manner if one exists in the same directory.
}
