% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparadox-package.R
\docType{package}
\name{Rparadox-package}
\alias{Rparadox}
\alias{Rparadox-package}
\title{Rparadox: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniil Popov \email{popov.daniil@gmail.com}

}
\keyword{internal}
