% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparadox-package.R
\docType{package}
\name{Rparadox-package}
\alias{Rparadox}
\alias{Rparadox-package}
\title{Rparadox: R Bindings for the Paradox Database Library}
\description{
Description: Provides an interface to the 'pxlib' C library for reading data from Paradox database files. This package allows users to import data from .db and .mb files into R data structures like tibbles, supporting various Paradox data types.
}
\author{
\strong{Maintainer}: Daniil Popov \email{popov.daniil@gmail.com}

}
\keyword{internal}
