% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rparadox-package.R
\docType{package}
\name{Rparadox-package}
\alias{Rparadox}
\alias{Rparadox-package}
\title{Rparadox: Read Paradox Database Files into R}
\description{
Provides a simple and efficient way to read data from Paradox database files (.db) directly into R as modern 'tibble' data frames. It uses the underlying 'pxlib' C library to handle the low-level file format details and provides a clean, user-friendly R interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/celebithil/Rparadox}
  \item Report bugs at \url{https://github.com/celebithil/Rparadox/issues}
}

}
\author{
\strong{Maintainer}: Daniil Popov \email{popov.daniil@gmail.com}

}
\keyword{internal}
